analytics/requirements.txt
kafka-python
influxdb-client
grpcio
protobuf
